refactor shared utility: ParallelMatrixVisitor
