+++ tpl-implementations/include/experimental/__p1673_bits/kokkos-kernels/blas2_matrix_rank_1_update.hpp
@@ -48,71 +48,9 @@
 #include "signal_kokkos_impl_called.hpp"
 #include "static_extent_match.hpp"
 #include "triangle.hpp"
+#include "parallel_matrix.hpp"
 
 namespace KokkosKernelsSTD {
-
-namespace Impl {
-
-// manages parallel execution of independent action
-// called like action(i, j) for each matrix element A(i, j)
-template <typename ExecSpace, typename MatrixType>
-class ParallelMatrixVisitor {
-public:
-  KOKKOS_INLINE_FUNCTION ParallelMatrixVisitor(ExecSpace &&exec_in, MatrixType A_in):
-    exec(exec_in), A(A_in), ext0(A.extent(0)), ext1(A.extent(1))
-  {}
-
-  template <typename ActionType>
-  KOKKOS_INLINE_FUNCTION
-  void for_each_matrix_element(ActionType action) {
-    if (ext0 > ext1) { // parallel rows
-      Kokkos::parallel_for(Kokkos::RangePolicy(exec, 0, ext0),
-        KOKKOS_LAMBDA(const auto i) {
-          using idx_type = std::remove_const_t<decltype(i)>;
-          for (idx_type j = 0; j < ext1; ++j) {
-            action(i, j);
-          }
-        });
-    } else { // parallel columns
-      Kokkos::parallel_for(Kokkos::RangePolicy(exec, 0, ext1),
-        KOKKOS_LAMBDA(const auto j) {
-          using idx_type = std::remove_const_t<decltype(j)>;
-          for (idx_type i = 0; i < ext0; ++i) {
-            action(i, j);
-          }
-        });
-    }
-    exec.fence();
-  }
-
-  template <typename ActionType>
-  void for_each_triangle_matrix_element(std::experimental::linalg::upper_triangle_t t, ActionType action) {
-    Kokkos::parallel_for(Kokkos::RangePolicy(exec, 0, ext1),
-      KOKKOS_LAMBDA(const auto j) {
-        using idx_type = std::remove_const_t<decltype(j)>;
-        for (idx_type i = 0; i <= j; ++i) {
-          action(i, j);
-        }
-      });
-    exec.fence();
-  }
-
-  template <typename ActionType>
-  void for_each_triangle_matrix_element(std::experimental::linalg::lower_triangle_t t, ActionType action) {
-    for_each_triangle_matrix_element(std::experimental::linalg::upper_triangle,
-        [action](const auto i, const auto j) {
-          action(j, i);
-      });
-  }
-
-private:
-  ExecSpace exec;
-  MatrixType A;
-  size_t ext0;
-  size_t ext1;
-};
-
-} // namespace Impl
 
 // Nonsymmetric non-conjugated rank-1 update
 // Performs BLAS xGER/xGERU (for real/complex types) A[i,j] += x[i] * y[j]
